# docs/swagger.yaml
definitions:
  handlers.errorResponse:
    properties:
      message:
        type: string
    type: object
  models.AdminSignIn:
    properties:
      fingerprint:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.AdminSignInResponse:
    properties:
      access:
        type: string
    type: object
  models.AuthResponse:
    properties:
      access:
        type: string
    type: object
  models.ContributeHandlerRequest:
    properties:
      author:
        type: string
      content:
        type: string
      source:
        type: string
      source_title:
        type: string
    type: object
...
  models.ModerationRejectToService:
    properties:
      moderationID:
        type: string
      reason:
        type: string
    type: object
  models.ModerationServiceResponse:
    properties:
      contributor_name:
        type: string
      moderation_id:
        type: string
      sent_at:
        type: string
    type: object
  models.ModerationTextDetails:
    properties:
      author:
        type: string
      content:
        type: string
      moderation_id:
        type: string
      source:
        type: string
      source_title:
        type: string
    type: object
...
  models.RefreshS:
    properties:
      fingerprint:
        type: string
    type: object
...
  models.SignIn:
    properties:
      email:
        type: string
      fingerprint:
        type: string
      password:
        type: string
    type: object
  models.SignUp:
    properties:
      email:
        type: string
      fingerprint:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
host: localhost:1001
info:
  contact: {}
  description: API for typeracer game clone. Typeracer popular game where users improve
    their typing skills in interactive format
  title: Typeracer Game clone API
  version: "2.0"
paths:
  /admin/auth/refresh:
    post:
      consumes:
      - application/json
      description: Admin can refresh their token
      operationId: refresh_admin
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AdminSignInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: refresh for admins
      tags:
      - admin
  /admin/auth/sign-in:
    post:
      consumes:
      - application/json
      description: This endpoint is used for user authentication.
      operationId: sign-in-admin
      parameters:
      - description: Sign In
        in: body
        name: models.AdminSignIn
        required: true
        schema:
          $ref: '#/definitions/models.AdminSignIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AdminSignInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Sign in for admins
      tags:
      - admin
  /admin/moderation/{moderation_id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific moderation item
      operationId: moderation-content
      parameters:
      - description: Moderation ID
        in: path
        name: moderation_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ModerationTextDetails'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Moderation details
      tags:
      - moderation
  /admin/moderation/all:
    get:
      consumes:
      - application/json
      description: This endpoint is used for see content in moderation
      operationId: moderation-all
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.ModerationServiceResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all text to see in pending mode
      tags:
      - moderation
  /admin/moderation/content/{moderation_id}/approve:
    post:
      consumes:
      - application/json
      description: Admin can approve this content and it appear in global text storage
      operationId: moderation-approve
      parameters:
      - description: Moderation ID
        in: path
        name: moderation_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Approve provided text content
      tags:
      - moderation
  /admin/moderation/content/{moderation_id}/reject:
    post:
      consumes:
      - application/json
      description: Admin can reject some content because of problem in content
      operationId: moderation-reject
      parameters:
      - description: Moderation ID
        in: path
        name: moderation_id
        required: true
        type: string
      - description: Reject
        in: body
        name: models.ModerationRejectToService
        required: true
        schema:
          $ref: '#/definitions/models.ModerationRejectToService'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Reject provided content
      tags:
      - moderation
  /api/auth/logout:
    delete:
      consumes:
      - application/json
      description: This endpoint is used for user logout.
      operationId: log-out
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - Bearer: []
      summary: Log out
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: This endpoint is used to refresh the endpoint token.
      operationId: refresh
      parameters:
      - description: Refresh
        in: body
        name: models.RefreshS
        required: true
        schema:
          $ref: '#/definitions/models.RefreshS'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Refresh token
      tags:
      - auth
  /api/auth/sign-in:
    post:
      consumes:
      - application/json
      description: This endpoint is used for user authentication.
      operationId: sign-in
      parameters:
      - description: Sign In
        in: body
        name: models.SignIn
        required: true
        schema:
          $ref: '#/definitions/models.SignIn'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Sign in
      tags:
      - auth
  /api/auth/sign-up:
    post:
      consumes:
      - application/json
      description: This endpoint is used for user registration.
      operationId: sign-up
      parameters:
      - description: Sign Up
        in: body
        name: models.SignUp
        required: true
        schema:
          $ref: '#/definitions/models.SignUp'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      summary: Sign up
      tags:
      - auth
  /content/contribute:
    post:
      consumes:
      - application/json
      description: Endpoint related to contribute text to the general text set
      operationId: contribute
      parameters:
      - description: Contribute
        in: body
        name: models.ContributeHandlerRequest
        required: true
        schema:
          $ref: '#/definitions/models.ContributeHandlerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Contribute text
      tags:
      - content
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"